Move protocol parsing from languageserver to metaserver
